Drop manual Pre-Depends on dpkg, raise debhelper b-d to ≥ 9.20141221, see #770226 + #770245
This simplifies debian/control quite a bit.

This commit should be still suitable for Jessie (if necessary) since
the mentioned debhelper version has just been unblocked for Jessie,
see https://bugs.debian.org/780016 -- On the other hand we're probably
on the safe side if we don't include it for any further Jessie
updates and just branch after the debian/5.0.7-5 tag.