Remove pre-#768937 legacy code from zsh{,-dev,-dbg,-doc}.postinst
Thanks to lintian for making me aware of this via its (experimental)
maintainer-script-may-use-dir_to_symlink_helper warning.

Actually these (now removed) code snippets did even the opposite of
what is defined in debian/*.maintscript for the according arch:any
packages. Either the failing rmdir saved us from the symlink being set
or the debian/*.maintscript generated snippets coming after the now
removed snippets fixed it just immediately again.

The only arch:all package in this list is zsh-doc which still has a
symlink /usr/share/doc/zsh-doc to zsh-common. But since this change
(in 5.0.2-3) predates Jessie as well as Trusty, we should be able to
safely remove that code now anyways.

If for some reason, this commit is starting troubles with zsh-doc
upgrades, you might want to add a file named
debian/zsh-doc.maintscript to the source package and put this single
line (without the leading spaces) into it:

  dir_to_symlink /usr/share/doc/zsh-doc zsh-common 5.0.2-3~

This line is untested though.