Install text-only FAQ into zsh-doc, too, for now. See #980903
debhelper currently has a bug when dh_installdocs is called with
-p<pkg> and the documents for a doc-base registration are split over
multiple packages (https://bugs.debian.org/980903) which is exactly
what I did in the previous commit.

So lets skip splitting the FAQ over two packages and install it only
into zsh-doc until #980903 is fixed in debhelper.

Once #980903 is fixed in debhelper, please do the following:

* Revert this commit.

* Add according Breaks and Replaces headers against zsh-doc to
  zsh-common.

* Add a build-dependency on debhelper >= the fixed version.

* Refer to #961757 in the debian/changelog entry (aka first line of
  a commit message).